Functions grammar structure is fixed
There are a few shift/reduce conflicts but it works just as expected.
The conflicts come from the implementation of the function grammar.
Since the function can include 'statements'..
And 'statements' get evaluated anywhere else but outside of that context,
 where the input start with FUNC... (or VAR as matter of fact)

I changed the grammar to have an evaluation rule.
The evaluation rule will determine whether or not a statement should get evaluated.
Eg. in a function declaration it does not...